+++ src/main/java/info/gehrels/voting/genderedElections/ElectionCalculationWithFemaleExclusivePositionsListener.java
@@ -3,9 +3,10 @@ package info.gehrels.voting.genderedElections;
 public interface ElectionCalculationWithFemaleExclusivePositionsListener {
 
 
-	void reducedNonFemaleExclusiveSeats(int numberOfOpenFemaleExclusiveSeats, int numberOfElectedFemaleExclusiveSeats,
-	                                    int numberOfOpenNonFemaleExclusiveSeats,
-	                                    int numberOfElectableNonFemaleExclusiveSeats);
+	void reducedNonFemaleExclusiveSeats(long numberOfOpenFemaleExclusiveSeats,
+	                                    long numberOfElectedFemaleExclusiveSeats,
+	                                    long numberOfOpenNonFemaleExclusiveSeats,
+	                                    long numberOfElectableNonFemaleExclusiveSeats);
 
 
 	void candidateNotQualified(GenderedCandidate candidate, String reason);
